refactor(#1134): split schema/mod.rs into cql_type_parser, udt_registry, schema_comparator
Behavior-preserving source split per epic #1116 doctrine. Move only:
- impl CqlType (parse, split_top_level_types, fixed_size, is_collection) + its
  parsing test -> schema/cql_type_parser.rs
- struct + impl UdtRegistry -> schema/udt_registry.rs (re-exported as
  schema::UdtRegistry to preserve the public path)
- TableSchema comparator methods -> schema/schema_comparator.rs

mod.rs keeps TableSchema, SchemaManager, data structs, is_udt_identifier, and
all re-exports. No logic/API changes.
