+++ cqlite-core/src/schema/mod.rs
@@ -12,6 +12,16 @@ pub mod json_exporter;
 pub mod parser;
 pub mod registry;
 
+// Concern submodules split out of this file (issue #1134, source-split
+// doctrine). They add inherent impls on the types defined here and own the
+// `UdtRegistry` type; the `pub use` below preserves the `schema::UdtRegistry`
+// public path.
+mod cql_type_parser;
+mod schema_comparator;
+mod udt_registry;
+
+pub use udt_registry::UdtRegistry;
+
 // Re-export aggregator components
 pub use aggregator::{
     AggregatorConfig, LoadErrorType, LoadResult, SchemaAggregator, SchemaLoadError,
@@ -53,7 +63,7 @@ use crate::error::{Error, Result};
 use crate::parser::header::SSTableHeader;
 use crate::parser::types::CqlTypeId;
 use crate::storage::StorageEngine;
-use crate::types::{ComparatorType, UdtTypeDef};
+use crate::types::UdtTypeDef;
 use crate::Config;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
@@ -232,343 +242,6 @@ pub enum CqlType {
 
     // Custom/Unknown
     Custom(String),
-}
-
-/// UDT Schema Registry for managing User Defined Type definitions
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct UdtRegistry {
-    /// Registered UDT type definitions by keyspace and type name
-    udts: HashMap<String, HashMap<String, UdtTypeDef>>,
-}
-
-impl UdtRegistry {
-    /// Create a new UDT registry
-    pub fn new() -> Self {
-        Self {
-            udts: HashMap::new(),
-        }
-    }
-
-    /// Create a new UDT registry with enhanced Cassandra 5.0 defaults
-    pub fn with_cassandra5_defaults() -> Self {
-        let mut registry = Self::new();
-        registry.load_cassandra5_system_udts();
-        registry
-    }
-
-    /// Register a UDT type definition
-    pub fn register_udt(&mut self, udt_def: UdtTypeDef) {
-        let keyspace_udts = self.udts.entry(udt_def.keyspace.clone()).or_default();
-        keyspace_udts.insert(udt_def.name.clone(), udt_def);
-    }
-
-    /// Get a UDT definition by keyspace and name
-    pub fn get_udt(&self, keyspace: &str, name: &str) -> Option<&UdtTypeDef> {
-        self.udts.get(keyspace)?.get(name)
-    }
-
-    /// Get all UDTs in a keyspace
-    pub fn get_keyspace_udts(&self, keyspace: &str) -> Option<&HashMap<String, UdtTypeDef>> {
-        self.udts.get(keyspace)
-    }
-
-    /// List all registered UDT names in a keyspace
-    pub fn list_udt_names(&self, keyspace: &str) -> Vec<&str> {
-        self.udts
-            .get(keyspace)
-            .map(|udts| udts.keys().map(|s| s.as_str()).collect())
-            .unwrap_or_default()
-    }
-
-    /// Check if a UDT is registered
-    pub fn contains_udt(&self, keyspace: &str, name: &str) -> bool {
-        self.udts
-            .get(keyspace)
-            .map(|udts| udts.contains_key(name))
-            .unwrap_or(false)
-    }
-
-    /// Remove a UDT definition
-    pub fn remove_udt(&mut self, keyspace: &str, name: &str) -> Option<UdtTypeDef> {
-        self.udts.get_mut(keyspace)?.remove(name)
-    }
-
-    /// Clear all UDTs in a keyspace
-    pub fn clear_keyspace(&mut self, keyspace: &str) {
-        self.udts.remove(keyspace);
-    }
-
-    /// Get total number of registered UDTs
-    pub fn total_udts(&self) -> usize {
-        self.udts.values().map(|udts| udts.len()).sum()
-    }
-
-    /// Load enhanced Cassandra 5.0 system UDTs with complex nested structures
-    fn load_cassandra5_system_udts(&mut self) {
-        // Enhanced address UDT for Cassandra 5.0 compatibility
-        let address_udt = UdtTypeDef::new("system".to_string(), "address".to_string())
-            .with_field("street".to_string(), CqlType::Text, true)
-            .with_field("street2".to_string(), CqlType::Text, true)
-            .with_field("city".to_string(), CqlType::Text, true)
-            .with_field("state".to_string(), CqlType::Text, true)
-            .with_field("zip_code".to_string(), CqlType::Text, true)
-            .with_field("country".to_string(), CqlType::Text, true)
-            .with_field(
-                "coordinates".to_string(),
-                CqlType::Tuple(vec![CqlType::Double, CqlType::Double]),
-                true,
-            );
-
-        self.register_udt(address_udt);
-
-        // Enhanced person UDT with collections and nested types
-        let person_udt = UdtTypeDef::new("system".to_string(), "person".to_string())
-            .with_field("id".to_string(), CqlType::Uuid, false)
-            .with_field("first_name".to_string(), CqlType::Text, false)
-            .with_field("last_name".to_string(), CqlType::Text, false)
-            .with_field("middle_name".to_string(), CqlType::Text, true)
-            .with_field("age".to_string(), CqlType::Int, true)
-            .with_field("email".to_string(), CqlType::Text, true)
-            .with_field(
-                "phone_numbers".to_string(),
-                CqlType::Set(Box::new(CqlType::Text)),
-                true,
-            )
-            .with_field(
-                "addresses".to_string(),
-                CqlType::List(Box::new(CqlType::Udt("address".to_string(), vec![]))),
-                true,
-            )
-            .with_field(
-                "metadata".to_string(),
-                CqlType::Map(Box::new(CqlType::Text), Box::new(CqlType::Text)),
-                true,
-            );
-
-        self.register_udt(person_udt);
-
-        // Contact info UDT for complex nested scenarios
-        let contact_info_udt = UdtTypeDef::new("system".to_string(), "contact_info".to_string())
-            .with_field(
-                "person".to_string(),
-                CqlType::Udt("person".to_string(), vec![]),
-                false,
-            )
-            .with_field(
-                "primary_address".to_string(),
-                CqlType::Udt("address".to_string(), vec![]),
-                true,
-            )
-            .with_field(
-                "emergency_contacts".to_string(),
-                CqlType::List(Box::new(CqlType::Udt("person".to_string(), vec![]))),
-                true,
-            )
-            .with_field("last_updated".to_string(), CqlType::Timestamp, true);
-
-        self.register_udt(contact_info_udt);
-    }
-
-    /// Resolve UDT with full dependency chain
-    pub fn resolve_udt_with_dependencies(
-        &self,
-        keyspace: &str,
-        name: &str,
-    ) -> crate::Result<&UdtTypeDef> {
-        let udt = self.get_udt(keyspace, name).ok_or_else(|| {
-            crate::Error::schema(format!(
-                "UDT '{}' not found in keyspace '{}'",
-                name, keyspace
-            ))
-        })?;
-
-        // Validate all field dependencies exist
-        for field in &udt.fields {
-            self.validate_field_type_dependencies(&field.field_type, keyspace)?;
-        }
-
-        Ok(udt)
-    }
-
-    /// Validate that all UDT field type dependencies exist in the registry
-    fn validate_field_type_dependencies(
-        &self,
-        field_type: &CqlType,
-        keyspace: &str,
-    ) -> crate::Result<()> {
-        match field_type {
-            CqlType::Udt(udt_name, _) => {
-                if !self.contains_udt(keyspace, udt_name) {
-                    return Err(crate::Error::schema(format!(
-                        "UDT dependency '{}' not found in keyspace '{}'",
-                        udt_name, keyspace
-                    )));
-                }
-            }
-            CqlType::List(inner) | CqlType::Set(inner) | CqlType::Frozen(inner) => {
-                self.validate_field_type_dependencies(inner, keyspace)?;
-            }
-            CqlType::Map(key_type, value_type) => {
-                self.validate_field_type_dependencies(key_type, keyspace)?;
-                self.validate_field_type_dependencies(value_type, keyspace)?;
-            }
-            CqlType::Tuple(field_types) => {
-                for tuple_field_type in field_types {
-                    self.validate_field_type_dependencies(tuple_field_type, keyspace)?;
-                }
-            }
-            _ => {} // Primitive types don't need validation
-        }
-        Ok(())
-    }
-
-    /// Get all UDTs that depend on a given UDT (for cascade operations)
-    pub fn get_dependent_udts(&self, keyspace: &str, udt_name: &str) -> Vec<&UdtTypeDef> {
-        let mut dependents = Vec::new();
-
-        if let Some(keyspace_udts) = self.udts.get(keyspace) {
-            for udt in keyspace_udts.values() {
-                if udt.name == udt_name {
-                    continue; // Skip self
-                }
-
-                // Check if this UDT depends on the target UDT
-                if self.udt_depends_on(udt, udt_name) {
-                    dependents.push(udt);
-                }
-            }
-        }
-
-        dependents
-    }
-
-    /// Check if a UDT depends on another UDT (recursively)
-    fn udt_depends_on(&self, udt: &UdtTypeDef, target_udt: &str) -> bool {
-        for field in &udt.fields {
-            if self.field_type_depends_on(&field.field_type, target_udt) {
-                return true;
-            }
-        }
-        false
-    }
-
-    /// Check if a field type depends on a UDT
-    #[allow(clippy::only_used_in_recursion)]
-    fn field_type_depends_on(&self, field_type: &CqlType, target_udt: &str) -> bool {
-        match field_type {
-            CqlType::Udt(udt_name, _) => udt_name == target_udt,
-            CqlType::List(inner) | CqlType::Set(inner) | CqlType::Frozen(inner) => {
-                self.field_type_depends_on(inner, target_udt)
-            }
-            CqlType::Map(key_type, value_type) => {
-                self.field_type_depends_on(key_type, target_udt)
-                    || self.field_type_depends_on(value_type, target_udt)
-            }
-            CqlType::Tuple(field_types) => field_types
-                .iter()
-                .any(|ft| self.field_type_depends_on(ft, target_udt)),
-            _ => false,
-        }
-    }
-
-    /// Register UDT with dependency validation
-    pub fn register_udt_with_validation(&mut self, udt_def: UdtTypeDef) -> crate::Result<()> {
-        // Validate dependencies exist
-        for field in &udt_def.fields {
-            self.validate_field_type_dependencies(&field.field_type, &udt_def.keyspace)?;
-        }
-
-        // Check for circular dependencies
-        if self.would_create_circular_dependency(&udt_def) {
-            return Err(crate::Error::schema(format!(
-                "Registering UDT '{}' would create circular dependency",
-                udt_def.name
-            )));
-        }
-
-        self.register_udt(udt_def);
-        Ok(())
-    }
-
-    /// Check if registering a UDT would create circular dependencies
-    fn would_create_circular_dependency(&self, udt_def: &UdtTypeDef) -> bool {
-        // This is complex - for now, just check direct self-reference
-        for field in &udt_def.fields {
-            if self.field_type_depends_on(&field.field_type, &udt_def.name) {
-                return true;
-            }
-        }
-        false
-    }
-
-    /// Export UDT definitions for debugging
-    pub fn export_definitions(&self, keyspace: &str) -> Vec<String> {
-        let mut definitions = Vec::new();
-
-        if let Some(keyspace_udts) = self.udts.get(keyspace) {
-            for udt in keyspace_udts.values() {
-                let mut def = format!("CREATE TYPE {}.{} (\n", keyspace, udt.name);
-
-                for (i, field) in udt.fields.iter().enumerate() {
-                    if i > 0 {
-                        def.push_str(",\n");
-                    }
-                    def.push_str(&format!(
-                        "  {} {}",
-                        field.name,
-                        self.format_cql_type(&field.field_type)
-                    ));
-                }
-
-                def.push_str("\n);");
-                definitions.push(def);
-            }
-        }
-
-        definitions
-    }
-
-    /// Format CQL type for CREATE TYPE statements
-    #[allow(clippy::only_used_in_recursion)]
-    fn format_cql_type(&self, cql_type: &CqlType) -> String {
-        match cql_type {
-            CqlType::Boolean => "boolean".to_string(),
-            CqlType::TinyInt => "tinyint".to_string(),
-            CqlType::SmallInt => "smallint".to_string(),
-            CqlType::Int => "int".to_string(),
-            CqlType::BigInt => "bigint".to_string(),
-            CqlType::Counter => "counter".to_string(),
-            CqlType::Float => "float".to_string(),
-            CqlType::Double => "double".to_string(),
-            CqlType::Text | CqlType::Varchar => "text".to_string(),
-            CqlType::Ascii => "ascii".to_string(),
-            CqlType::Blob => "blob".to_string(),
-            CqlType::Timestamp => "timestamp".to_string(),
-            CqlType::Date => "date".to_string(),
-            CqlType::Time => "time".to_string(),
-            CqlType::Uuid => "uuid".to_string(),
-            CqlType::TimeUuid => "timeuuid".to_string(),
-            CqlType::Inet => "inet".to_string(),
-            CqlType::Duration => "duration".to_string(),
-            CqlType::Varint => "varint".to_string(),
-            CqlType::Decimal => "decimal".to_string(),
-            CqlType::List(inner) => format!("list<{}>", self.format_cql_type(inner)),
-            CqlType::Set(inner) => format!("set<{}>", self.format_cql_type(inner)),
-            CqlType::Map(key, value) => format!(
-                "map<{}, {}>",
-                self.format_cql_type(key),
-                self.format_cql_type(value)
-            ),
-            CqlType::Udt(name, _) => name.clone(),
-            CqlType::Tuple(types) => {
-                let type_strs: Vec<String> =
-                    types.iter().map(|t| self.format_cql_type(t)).collect();
-                format!("tuple<{}>", type_strs.join(", "))
-            }
-            CqlType::Frozen(inner) => format!("frozen<{}>", self.format_cql_type(inner)),
-            CqlType::Custom(name) => name.clone(),
-        }
-    }
 }
 
 /// Whether a de-prefixed `Custom` type name is a plausible UDT reference.
@@ -1008,136 +681,6 @@ impl TableSchema {
         keys
     }
 
-    /// Get ComparatorType for a specific column
-    pub fn get_column_comparator(&self, column_name: &str) -> Result<ComparatorType> {
-        let column = self
-            .get_column(column_name)
-            .ok_or_else(|| Error::Schema(format!("Column '{}' not found", column_name)))?;
-
-        let cql_type = CqlType::parse(&column.data_type)?;
-        ComparatorType::from_cql_type(&cql_type)
-    }
-
-    /// Get ComparatorTypes for all columns
-    pub fn get_all_comparators(&self) -> Result<HashMap<String, ComparatorType>> {
-        let mut comparators = HashMap::new();
-
-        for column in &self.columns {
-            let cql_type = CqlType::parse(&column.data_type)?;
-            let comparator = ComparatorType::from_cql_type(&cql_type)?;
-            comparators.insert(column.name.clone(), comparator);
-        }
-
-        Ok(comparators)
-    }
-
-    /// Get ComparatorTypes for partition key columns in order
-    pub fn get_partition_key_comparators(&self) -> Result<Vec<ComparatorType>> {
-        let mut comparators = Vec::new();
-        let ordered_keys = self.ordered_partition_keys();
-
-        for key_column in ordered_keys {
-            let cql_type = CqlType::parse(&key_column.data_type)?;
-            let comparator = ComparatorType::from_cql_type(&cql_type)?;
-            comparators.push(comparator);
-        }
-
-        Ok(comparators)
-    }
-
-    /// Get ComparatorTypes for clustering key columns in order
-    pub fn get_clustering_key_comparators(&self) -> Result<Vec<ComparatorType>> {
-        let mut comparators = Vec::new();
-        let ordered_keys = self.ordered_clustering_keys();
-
-        for key_column in ordered_keys {
-            let cql_type = CqlType::parse(&key_column.data_type)?;
-            let comparator = ComparatorType::from_cql_type(&cql_type)?;
-            comparators.push(comparator);
-        }
-
-        Ok(comparators)
-    }
-
-    /// Check if a column type is compatible with an expected type
-    pub fn is_column_type_compatible(
-        &self,
-        column_name: &str,
-        expected_type: &str,
-    ) -> Result<bool> {
-        let column_comparator = self.get_column_comparator(column_name)?;
-        let expected_cql_type = CqlType::parse(expected_type)?;
-        let expected_comparator = ComparatorType::from_cql_type(&expected_cql_type)?;
-
-        Ok(self.comparators_are_compatible(&column_comparator, &expected_comparator))
-    }
-
-    /// Check if two ComparatorTypes are compatible (helper method)
-    #[allow(clippy::only_used_in_recursion)]
-    fn comparators_are_compatible(&self, left: &ComparatorType, right: &ComparatorType) -> bool {
-        match (left, right) {
-            // Exact matches
-            (ComparatorType::Boolean, ComparatorType::Boolean) => true,
-            (ComparatorType::TinyInt, ComparatorType::TinyInt) => true,
-            (ComparatorType::SmallInt, ComparatorType::SmallInt) => true,
-            (ComparatorType::Int, ComparatorType::Int) => true,
-            (ComparatorType::BigInt, ComparatorType::BigInt) => true,
-            (ComparatorType::Float32, ComparatorType::Float32) => true,
-            (ComparatorType::Float, ComparatorType::Float) => true,
-            (ComparatorType::Text, ComparatorType::Text) => true,
-            (ComparatorType::Blob, ComparatorType::Blob) => true,
-            (ComparatorType::Timestamp, ComparatorType::Timestamp) => true,
-            (ComparatorType::Uuid, ComparatorType::Uuid) => true,
-            (ComparatorType::Json, ComparatorType::Json) => true,
-
-            // Collection types
-            (ComparatorType::List(l_elem), ComparatorType::List(r_elem)) => {
-                self.comparators_are_compatible(l_elem, r_elem)
-            }
-            (ComparatorType::Set(l_elem), ComparatorType::Set(r_elem)) => {
-                self.comparators_are_compatible(l_elem, r_elem)
-            }
-            (ComparatorType::Map(l_key, l_val), ComparatorType::Map(r_key, r_val)) => {
-                self.comparators_are_compatible(l_key, r_key)
-                    && self.comparators_are_compatible(l_val, r_val)
-            }
-
-            // Tuple types
-            (ComparatorType::Tuple(l_fields), ComparatorType::Tuple(r_fields)) => {
-                l_fields.len() == r_fields.len()
-                    && l_fields
-                        .iter()
-                        .zip(r_fields.iter())
-                        .all(|(l, r)| self.comparators_are_compatible(l, r))
-            }
-
-            // UDT types
-            (
-                ComparatorType::Udt {
-                    type_name: l_name,
-                    keyspace: l_ks,
-                    ..
-                },
-                ComparatorType::Udt {
-                    type_name: r_name,
-                    keyspace: r_ks,
-                    ..
-                },
-            ) => l_name == r_name && l_ks == r_ks,
-
-            // Frozen types
-            (ComparatorType::Frozen(l_inner), ComparatorType::Frozen(r_inner)) => {
-                self.comparators_are_compatible(l_inner, r_inner)
-            }
-
-            // Custom types
-            (ComparatorType::Custom(l_name), ComparatorType::Custom(r_name)) => l_name == r_name,
-
-            // No other combinations are compatible
-            _ => false,
-        }
-    }
-
     /// Create a minimal test schema (for testing only)
     #[cfg(test)]
     pub fn new_for_testing(keyspace: &str, table: &str) -> Self {
@@ -1160,214 +703,6 @@ impl TableSchema {
             comments: HashMap::new(),
             dropped_columns: HashMap::new(),
         }
-    }
-}
-
-impl CqlType {
-    fn split_top_level_types(type_str: &str) -> Result<Vec<&str>> {
-        let mut parts = Vec::new();
-        let mut depth = 0usize;
-        let mut start = 0usize;
-
-        for (index, ch) in type_str.char_indices() {
-            match ch {
-                '<' => depth += 1,
-                '>' => {
-                    if depth == 0 {
-                        return Err(Error::schema(format!(
-                            "Invalid nested type syntax: {}",
-                            type_str
-                        )));
-                    }
-                    depth -= 1;
-                }
-                ',' if depth == 0 => {
-                    parts.push(type_str[start..index].trim());
-                    start = index + ch.len_utf8();
-                }
-                _ => {}
-            }
-        }
-
-        if depth != 0 {
-            return Err(Error::schema(format!(
-                "Unbalanced nested type syntax: {}",
-                type_str
-            )));
-        }
-
-        parts.push(type_str[start..].trim());
-        Ok(parts.into_iter().filter(|part| !part.is_empty()).collect())
-    }
-
-    /// Parse CQL type string into structured type
-    pub fn parse(type_str: &str) -> Result<Self> {
-        let type_str = type_str.trim();
-
-        // CQL type keywords are case-insensitive (`SET<TEXT>` == `set<text>`),
-        // so match collection/frozen/tuple prefixes case-insensitively. Matching
-        // only lowercase here previously left uppercase collections to fall
-        // through to a bare `Custom("SET<TEXT>")`, which both broke type-aware
-        // handling and confused UDT-reference validation (roborev job 51).
-        fn strip_prefix_ci<'a>(s: &'a str, prefix: &str) -> Option<&'a str> {
-            s.get(..prefix.len())
-                .filter(|head| head.eq_ignore_ascii_case(prefix))
-                .map(|_| &s[prefix.len()..])
-        }
-
-        // Handle frozen types
-        if let Some(inner) = strip_prefix_ci(type_str, "frozen<") {
-            if let Some(inner) = inner.strip_suffix('>') {
-                return Ok(CqlType::Frozen(Box::new(Self::parse(inner)?)));
-            }
-        }
-
-        // Handle collection types
-        if let Some(inner) = strip_prefix_ci(type_str, "list<") {
-            if let Some(inner) = inner.strip_suffix('>') {
-                return Ok(CqlType::List(Box::new(Self::parse(inner)?)));
-            }
-        }
-
-        if let Some(inner) = strip_prefix_ci(type_str, "set<") {
-            if let Some(inner) = inner.strip_suffix('>') {
-                return Ok(CqlType::Set(Box::new(Self::parse(inner)?)));
-            }
-        }
-
-        if let Some(inner) = strip_prefix_ci(type_str, "map<") {
-            if let Some(inner) = inner.strip_suffix('>') {
-                let parts = Self::split_top_level_types(inner)?;
-                if parts.len() != 2 {
-                    return Err(Error::schema(format!("Invalid map type: {}", type_str)));
-                }
-                return Ok(CqlType::Map(
-                    Box::new(Self::parse(parts[0].trim())?),
-                    Box::new(Self::parse(parts[1].trim())?),
-                ));
-            }
-        }
-
-        // Handle tuple types
-        if let Some(inner) = strip_prefix_ci(type_str, "tuple<") {
-            if let Some(inner) = inner.strip_suffix('>') {
-                let parts = Self::split_top_level_types(inner)?;
-                let mut types = Vec::new();
-                for part in parts {
-                    types.push(Self::parse(part.trim())?);
-                }
-                return Ok(CqlType::Tuple(types));
-            }
-        }
-
-        // Handle UDT types - format: udt_name or keyspace.udt_name
-        // But first check if it's not a primitive type in uppercase
-        let lowercase_type = type_str.to_lowercase();
-        let is_primitive = matches!(
-            lowercase_type.as_str(),
-            "boolean"
-                | "bool"
-                | "tinyint"
-                | "smallint"
-                | "int"
-                | "integer"
-                | "bigint"
-                | "long"
-                | "counter"
-                | "float"
-                | "double"
-                | "decimal"
-                | "text"
-                | "varchar"
-                | "ascii"
-                | "blob"
-                | "timestamp"
-                | "date"
-                | "time"
-                | "uuid"
-                | "timeuuid"
-                | "inet"
-                | "duration"
-        );
-
-        if !is_primitive
-            && type_str
-                .chars()
-                .all(|c| c.is_alphanumeric() || c == '_' || c == '.')
-            && !type_str.chars().all(|c| c.is_ascii_lowercase())
-        {
-            // This might be a UDT name - store as custom type for now
-            // Full validation requires UDT registry context
-            return Ok(CqlType::Custom(format!("udt:{}", type_str)));
-        }
-
-        // Primitive types
-        match type_str.to_lowercase().as_str() {
-            "boolean" | "bool" => Ok(CqlType::Boolean),
-            "tinyint" => Ok(CqlType::TinyInt),
-            "smallint" => Ok(CqlType::SmallInt),
-            "int" | "integer" => Ok(CqlType::Int),
-            "bigint" | "long" => Ok(CqlType::BigInt),
-            "counter" => Ok(CqlType::Counter),
-            "float" => Ok(CqlType::Float),
-            "double" => Ok(CqlType::Double),
-            "decimal" => Ok(CqlType::Decimal),
-            "text" | "varchar" => Ok(CqlType::Text),
-            "ascii" => Ok(CqlType::Ascii),
-            "blob" => Ok(CqlType::Blob),
-            "timestamp" => Ok(CqlType::Timestamp),
-            "date" => Ok(CqlType::Date),
-            "time" => Ok(CqlType::Time),
-            "uuid" => Ok(CqlType::Uuid),
-            "timeuuid" => Ok(CqlType::TimeUuid),
-            "inet" => Ok(CqlType::Inet),
-            "duration" => Ok(CqlType::Duration),
-            "varint" => Ok(CqlType::Varint),
-            _ => Ok(CqlType::Custom(type_str.to_string())),
-        }
-    }
-
-    /// Get the expected byte size for fixed-size types
-    pub fn fixed_size(&self) -> Option<usize> {
-        match self {
-            CqlType::Boolean => Some(1),
-            CqlType::TinyInt => Some(1),
-            CqlType::SmallInt => Some(2),
-            CqlType::Int => Some(4),
-            CqlType::BigInt => Some(8),
-            CqlType::Counter => Some(8),
-            CqlType::Float => Some(4),
-            CqlType::Double => Some(8),
-            CqlType::Timestamp => Some(8),
-            CqlType::Date => Some(4),
-            CqlType::Time => Some(8),
-            CqlType::Uuid | CqlType::TimeUuid => Some(16),
-            CqlType::Inet => Some(16), // IPv6, IPv4 is variable
-            // Variable size types
-            CqlType::Text
-            | CqlType::Ascii
-            | CqlType::Varchar
-            | CqlType::Blob
-            | CqlType::Decimal
-            | CqlType::Duration
-            | CqlType::Varint => None,
-            // Collections and complex types are variable
-            CqlType::List(_)
-            | CqlType::Set(_)
-            | CqlType::Map(_, _)
-            | CqlType::Tuple(_)
-            | CqlType::Udt(_, _) => None,
-            CqlType::Frozen(inner) => inner.fixed_size(),
-            CqlType::Custom(_) => None,
-        }
-    }
-
-    /// Check if this type is a collection
-    pub fn is_collection(&self) -> bool {
-        matches!(
-            self,
-            CqlType::List(_) | CqlType::Set(_) | CqlType::Map(_, _)
-        )
     }
 }
 
@@ -1672,38 +1007,6 @@ mod tests {
         assert_eq!(schema.table, "users");
         assert_eq!(schema.partition_keys.len(), 1);
         assert_eq!(schema.columns.len(), 2);
-    }
-
-    #[test]
-    fn test_cql_type_parsing() {
-        assert_eq!(CqlType::parse("text").unwrap(), CqlType::Text);
-        assert_eq!(CqlType::parse("bigint").unwrap(), CqlType::BigInt);
-
-        match CqlType::parse("list<int>").unwrap() {
-            CqlType::List(inner) => assert_eq!(*inner, CqlType::Int),
-            _ => panic!("Expected List type"),
-        }
-
-        match CqlType::parse("map<text, bigint>").unwrap() {
-            CqlType::Map(key, value) => {
-                assert_eq!(*key, CqlType::Text);
-                assert_eq!(*value, CqlType::BigInt);
-            }
-            _ => panic!("Expected Map type"),
-        }
-
-        match CqlType::parse("tuple<text, list<int>, map<text, text>>").unwrap() {
-            CqlType::Tuple(fields) => {
-                assert_eq!(fields.len(), 3);
-                assert_eq!(fields[0], CqlType::Text);
-                assert_eq!(fields[1], CqlType::List(Box::new(CqlType::Int)));
-                assert_eq!(
-                    fields[2],
-                    CqlType::Map(Box::new(CqlType::Text), Box::new(CqlType::Text))
-                );
-            }
-            _ => panic!("Expected Tuple type"),
-        }
     }
 
     #[test]
